Added module roles and permissions settings to the admin module.
Changed data type of sort_order field in cms_permissions for postgres and sqlite.

Added links to the roles and permissions settings in the nav.

Fixed a bug in TableForm where a row without a description would not show the third column as blank even though other rows have descriptions causing gaps to appear.

Removed comments in an assert in the Table object that were causing it to always fail.

Removed the togglemodules page.

+++ scripts/install/database/sqlite.sql
@@ -135,7 +135,7 @@ CREATE TABLE cms_permissions (
   display_name varchar(100) NOT NULL,
   description varchar(255),
   module_id integer NOT NULL,
-  sort_order smallint NOT NULL,
+  sort_order integer NOT NULL
   FOREIGN KEY (module_id) REFERENCES cms_modules (module_id) ON DELETE CASCADE ON UPDATE CASCADE
 );
